// frontend/src/app/core/generated/circabc/model/groupDeletionRequestApproval.ts
/**
 * CIRCABC REST API
 *
 * Contact: DIGIT-CIRCABC-SUPPORT@ec.europa.eu
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface GroupDeletionRequestApproval { 
    id: number;
    /**
     * 1 == agreed -1 == refused 0 == waiting approval 
     */
    agreement: number;
    argument?: string;
}